+++ x/gov/abci.go
@@ -1,10 +1,12 @@
 package gov
 
 import (
+	"errors"
 	"fmt"
 	"time"
 
 	"cosmossdk.io/collections"
+	"cosmossdk.io/log"
 
 	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/telemetry"
@@ -25,10 +27,26 @@ func EndBlocker(ctx sdk.Context, keeper *keeper.Keeper) error {
 	err := keeper.InactiveProposalsQueue.Walk(ctx, rng, func(key collections.Pair[time.Time, uint64], _ uint64) (bool, error) {
 		proposal, err := keeper.Proposals.Get(ctx, key.K2())
 		if err != nil {
+			// if the proposal has an encoding error, this means it cannot be processed by x/gov
+			// this could be due to some types missing their registration
+			// instead of returning an error (i.e, halting the chain), we fail the proposal
+			if errors.Is(err, collections.ErrEncoding) {
+				proposal.Id = key.K2()
+				if err := failUnsupportedProposal(logger, ctx, keeper, proposal, err.Error(), false); err != nil {
+					return false, err
+				}
+
+				if err = keeper.DeleteProposal(ctx, proposal.Id); err != nil {
+					return false, err
+				}
+
+				return false, nil
+			}
+
 			return false, err
 		}
-		err = keeper.DeleteProposal(ctx, proposal.Id)
-		if err != nil {
+
+		if err = keeper.DeleteProposal(ctx, proposal.Id); err != nil {
 			return false, err
 		}
 
@@ -77,6 +95,22 @@ func EndBlocker(ctx sdk.Context, keeper *keeper.Keeper) error {
 	err = keeper.ActiveProposalsQueue.Walk(ctx, rng, func(key collections.Pair[time.Time, uint64], _ uint64) (bool, error) {
 		proposal, err := keeper.Proposals.Get(ctx, key.K2())
 		if err != nil {
+			// if the proposal has an encoding error, this means it cannot be processed by x/gov
+			// this could be due to some types missing their registration
+			// instead of returning an error (i.e, halting the chain), we fail the proposal
+			if errors.Is(err, collections.ErrEncoding) {
+				proposal.Id = key.K2()
+				if err := failUnsupportedProposal(logger, ctx, keeper, proposal, err.Error(), true); err != nil {
+					return false, err
+				}
+
+				if err = keeper.ActiveProposalsQueue.Remove(ctx, collections.Join(*proposal.VotingEndTime, proposal.Id)); err != nil {
+					return false, err
+				}
+
+				return false, nil
+			}
+
 			return false, err
 		}
 
@@ -97,14 +131,12 @@ func EndBlocker(ctx sdk.Context, keeper *keeper.Keeper) error {
 			} else {
 				err = keeper.RefundAndDeleteDeposits(ctx, proposal.Id)
 			}
-
 			if err != nil {
 				return false, err
 			}
 		}
 
-		err = keeper.ActiveProposalsQueue.Remove(ctx, collections.Join(*proposal.VotingEndTime, proposal.Id))
-		if err != nil {
+		if err = keeper.ActiveProposalsQueue.Remove(ctx, collections.Join(*proposal.VotingEndTime, proposal.Id)); err != nil {
 			return false, err
 		}
 
@@ -234,4 +266,49 @@ func safeExecuteHandler(ctx sdk.Context, msg sdk.Msg, handler baseapp.MsgService
 	}()
 	res, err = handler(ctx, msg)
 	return
+}
+
+// failUnsupportedProposal fails a proposal that cannot be processed by gov
+func failUnsupportedProposal(
+	logger log.Logger,
+	ctx sdk.Context,
+	keeper *keeper.Keeper,
+	proposal v1.Proposal,
+	errMsg string,
+	active bool,
+) error {
+	proposal.Status = v1.StatusFailed
+	proposal.FailedReason = fmt.Sprintf("proposal failed because it cannot be processed by gov: %s", errMsg)
+	proposal.Messages = nil // clear out the messages
+
+	if err := keeper.SetProposal(ctx, proposal); err != nil {
+		return err
+	}
+
+	if err := keeper.RefundAndDeleteDeposits(ctx, proposal.Id); err != nil {
+		return err
+	}
+
+	eventType := types.EventTypeInactiveProposal
+	if active {
+		eventType = types.EventTypeActiveProposal
+	}
+
+	ctx.EventManager().EmitEvent(
+		sdk.NewEvent(
+			eventType,
+			sdk.NewAttribute(types.AttributeKeyProposalID, fmt.Sprintf("%d", proposal.Id)),
+			sdk.NewAttribute(types.AttributeKeyProposalResult, types.AttributeValueProposalFailed),
+		),
+	)
+
+	logger.Info(
+		"proposal failed to decode; deleted",
+		"proposal", proposal.Id,
+		"expedited", proposal.Expedited,
+		"title", proposal.Title,
+		"results", errMsg,
+	)
+
+	return nil
 }